mgr/dashboard: extract monkey patches for cherrypy out
so they are less distracting to the dashboard developers. and probably
less scaring.

(cherry picked from commit 01cb398019e4e0050eea24b7be1a28186d042919)
